// src/transaction/enums/transaction.enum.ts
export enum TransactionType {
  CREDIT = 'CREDIT',
  DEBIT = 'DEBIT',
  BILL_PAYMENT = 'BILL_PAYMENT',
}

export enum TransactionStatus {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  REVERSED = 'REVERSED',
  REVERSAL_PENDING = 'REVERSAL_PENDING',
}